Fixing CPF validation, and CEP validation

+++ resources/user.py
@@ -2,6 +2,8 @@ from fastapi import (APIRouter, HTTPException, Request)
 from config.config import parse_configuration
 from controller.bases.user import (User, UpdateUser)
 from controller.database import DatabaseController
+from utils.valid_cpf import validate
+from utils.consult_cep import consult
 
 config = parse_configuration()
 
@@ -11,13 +13,24 @@ db = DatabaseController(config['postgresql'])
 
 
 @router.get('/users/user/', status_code=200)
-def consulta_user(cpf: str):
+def get_user(cpf: str):
+    if not validate(cpf):
+        return HTTPException(
+            500,
+            "Invalid CPF"
+        )
     if not cpf:
         return HTTPException(
             400,
             "User is null"
         )
-    return db.get_user(cpf)
+    data = db.get_user(cpf)
+    if len(data) < 1:
+        return HTTPException(
+            404,
+            "User not found"
+        )
+    return data
 
 
 @router.get('/users/', status_code=200)
@@ -27,6 +40,24 @@ def get_users():
 
 @router.post('/users/user/', status_code=201)
 def insert_user(user: User):
+    if not validate(user.cpf):
+        return HTTPException(
+            500,
+            "Invalid CPF"
+        )
+
+    cep_valid = consult(user.cep)
+    cep_json = cep_valid[1]
+    if not cep_valid[0]:
+        return HTTPException(
+            500,
+            "Invalid CPF"
+        )
+
+    user.estado = cep_json.get('uf')
+    user.cidade = cep_json.get('localidade')
+    user.rua = cep_json.get('logradouro')
+
     success = db.add_user(user)
     if not success:
         return HTTPException(
@@ -38,6 +69,18 @@ def insert_user(user: User):
 
 @router.put("/users/user/")
 def update_user(user: UpdateUser):
+    if not validate(user.cpf):
+        return HTTPException(
+            500,
+            "Invalid CPF"
+        )
+
+    if len(db.get_user(user.cpf)) < 1:
+        return HTTPException(
+            404,
+            "User not found"
+        )
+
     success = db.update_user(user)
     if not success:
         return HTTPException(
@@ -49,6 +92,18 @@ def update_user(user: UpdateUser):
 
 @router.delete("/users/user/")
 def remove_user(cpf: str):
+    if not validate(cpf):
+        return HTTPException(
+            500,
+            "Invalid CPF"
+        )
+
+    if len(db.get_user(cpf)) < 1:
+        return HTTPException(
+            404,
+            "User not found"
+        )
+
     success = db.delete_user(cpf)
     if not success:
         return HTTPException(
